Simplified API by using shared pointers

See merge request loco-3d/crocoddyl!245

+++ src/core/activations/quadratic-barrier.cpp
@@ -11,13 +11,13 @@
 namespace crocoddyl {
 
 ActivationModelQuadraticBarrier::ActivationModelQuadraticBarrier(const ActivationBounds& bounds)
-    : ActivationModelAbstract((unsigned int)bounds.lb.size()), bounds_(bounds) {}
+    : ActivationModelAbstract(bounds.lb.size()), bounds_(bounds) {}
 
 ActivationModelQuadraticBarrier::~ActivationModelQuadraticBarrier() {}
 
 void ActivationModelQuadraticBarrier::calc(const boost::shared_ptr<ActivationDataAbstract>& data,
                                            const Eigen::Ref<const Eigen::VectorXd>& r) {
-  assert(r.size() == nr_ && "r has wrong dimension");
+  assert(static_cast<std::size_t>(r.size()) == nr_ && "r has wrong dimension");
   boost::shared_ptr<ActivationDataQuadraticBarrier> d =
       boost::static_pointer_cast<ActivationDataQuadraticBarrier>(data);
 
@@ -28,7 +28,7 @@ void ActivationModelQuadraticBarrier::calc(const boost::shared_ptr<ActivationDat
 
 void ActivationModelQuadraticBarrier::calcDiff(const boost::shared_ptr<ActivationDataAbstract>& data,
                                                const Eigen::Ref<const Eigen::VectorXd>& r, const bool& recalc) {
-  assert(r.size() == nr_ && "r has wrong dimension");
+  assert(static_cast<std::size_t>(r.size()) == nr_ && "r has wrong dimension");
   if (recalc) {
     calc(data, r);
   }
